requirements.txt
# ACHE SUCATAS V14 - Dependências Python
# Instalar: pip install -r requirements.txt

# PDF Parsing
pdfplumber>=0.10.0

# Data manipulation
pandas>=2.0.0
openpyxl>=3.1.0

# Supabase (PostgreSQL + Storage)
supabase>=2.0.0

# Data validation (Miner V11)
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Async (Miner V11)
aiohttp>=3.9.0
aiofiles>=23.0.0

# DOCX parsing (opcional)
python-docx>=1.0.0

# Dashboard
streamlit>=1.30.0

# Webhook Server (Asaas Integration)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
